Slime AI
A basic slime AI. It has three types of jumps: high to the side, short to the side, and short straight up. When the player is outside of its detection radius it will jump around at random intervals and in random directions, favoring short jumps. When the player is near it will attempt to jump towards the player, using high jumps if the player is far and short jumps if they are near. The slime also tracks the player with its eyes and gets mean when it can see the player. Its cute.